+++ queries/useUserQueries.ts
@@ -1,9 +1,11 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
+import signout from '@apis/oauth/signout';
 import deleteProfileImage from '@apis/user/mypage/deleteProfileImage';
 import getUserDetail from '@apis/user/mypage/getUserDetail';
 import updateUserProfile from '@apis/user/mypage/updateUserProfile';
 import checkBlobId from '@apis/user/sign/checkBlobId';
+import { useUserStore } from '@stores/userStore';
 
 import { users } from './keys/userQueryKeys';
 
@@ -58,4 +60,31 @@ export function useDeleteProfileImage(blobId: string) {
       console.error('삭제실패ㅜㅜ:', error);
     },
   });
+}
+
+// 로그아웃
+export function useSignout() {
+  const queryClient = useQueryClient();
+  const { signout: logout } = useUserStore();
+
+  return useMutation({
+    mutationFn: signout,
+    onSuccess: () => {
+      console.log('로그아웃 성공');
+
+      queryClient.invalidateQueries({ queryKey: users.all.queryKey });
+
+      logout();
+      deleteCookies();
+      localStorage.clear();
+    },
+    onError: (error) => {
+      console.error('로그아웃 실패ㅜㅜ:', error);
+    },
+  });
+}
+
+function deleteCookies() {
+  document.cookie = `accessToken=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;`;
+  document.cookie = `refreshToken=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;`;
 }
